\name{HypoGym-Package}
\alias{HypoGym-Package}
\docType{package}
\title{
\packageTitle{HypoGym}
}
\description{This package contains a set of datasets for a hypothetical small gym and a series of functions. These functions serve to collect the data over time and display them in both numeric and graphical form. These functions can be
applied to many different datasets.
}

\author{Matthew Newbold

\maintainer: \mnewbold812{HypoGym}
}


